+++ app/src/main/java/app/pablopatarca/thenotestaker/domain/NotesRepository.kt
@@ -6,7 +6,7 @@ interface NotesRepository {
 
     fun getNotes(): Flow<List<Note>>
 
-    suspend fun getNoteById(id: Int): Note?
+    suspend fun getNoteById(id: Long): Note?
 
     suspend fun insert(note: Note)
